requirements.txt
django==2.0.5
django-cors-headers
django-background-tasks
